requirements.txt
Flask==2.3.2
langchain==0.1.3
langchain-community==0.0.15
snowflake-sqlalchemy
tiktoken
langchain_openai
python-dotenv==1.0.0
Flask-CORS